runtime: separate message value types
