+++ internal/runtime/funcs/get_dict_value_by_key_test.go
@@ -13,7 +13,7 @@ func TestGetDictValueTypedMissSendsError(t *testing.T) {
 	t.Parallel()
 
 	io, inputs, outputs := newIO([]string{"dict", "key"}, []string{"res", "err"})
-	handler, err := (dictGetByKey{}).Create(io, nil)
+	handler, err := (getDictValueByKey{}).Create(io, nil)
 	if err != nil {
 		t.Fatalf("Create returned error: %v", err)
 	}
@@ -46,7 +46,7 @@ func TestGetDictValueSendsTypedValue(t *testing.T) {
 	t.Parallel()
 
 	io, inputs, outputs := newIO([]string{"dict", "key"}, []string{"res", "err"})
-	handler, err := (dictGetByKey{}).Create(io, nil)
+	handler, err := (getDictValueByKey{}).Create(io, nil)
 	if err != nil {
 		t.Fatalf("Create returned error: %v", err)
 	}
