+++ internal/runtime/messages/string.go
@@ -1,0 +1,32 @@
+package messages
+
+import "encoding/json"
+
+// --- STRING ---
+//
+//nolint:godoclint // TODO(strict-lint phase 1): temporary suppression; remove after strict cleanup.
+type StringMsg struct {
+	internalMsg
+	v string
+}
+
+func (msg StringMsg) Str() string { return msg.v }
+
+func (msg StringMsg) String() string { return msg.v }
+
+func (msg StringMsg) MarshalJSON() ([]byte, error) {
+	//nolint:wrapcheck // TODO(strict-lint phase 1): temporary suppression; remove after strict cleanup.
+	return json.Marshal(msg.String())
+}
+
+func NewStringMsg(s string) StringMsg {
+	return StringMsg{
+		internalMsg: internalMsg{},
+		v:           s,
+	}
+}
+
+func equalString(left StringMsg, right Msg) bool {
+	rightTyped, ok := right.(StringMsg)
+	return ok && left.v == rightTyped.v
+}
