+++ std/builtin/core.neva
@@ -71,7 +71,7 @@ pub def Union<T Type>(data Type, tag T) (res T)
 // Get retrieves a value from a dictionary using the provided key.
 // It has two inports: 'dict' for the dictionary and 'key' for the lookup key.
 // Sends result to 'res' outport, or error to 'err' if key not found.
-#extern(dict_get_by_key)
+#extern(get_dict_value_by_key)
 pub def Get<T>(dict dict<T>, key T) (res T, err error)
 
 // Pass is used to pass data through. It receives data and sends it to outport.
